chore: Move deck methods to deck.go
moved deck methods to deck.go to clean up main.go. other changes are auto formatting "go fmt" command
